cleanly decouple api and no-api firecracker flavors
We can now cleanly decouple api and no-api firecracker flavors.

The controller had its event loop outside to allow external
handling of control events.

Add a `run()` function that runs the event loop internally
and takes a Trait object that defines a control event handler.
This way, all users of the controller can simply implement
their specific control event handler without having to implement
the whole event loop.

Use this functionality in both our api and no-api microVM flavors.
